# Optimize weekly_force_merge_stats query performance