## Summary
- Optimize the weekly_force_merge_stats query performance by ~25%
- Maintain identical query output and behavior

## Optimizations
1. Reordered time filters for earlier filtering
2. Used startsWith() instead of LIKE for URL prefix matching
3. Added pre-filtering of tables before JOIN operations
4. Simplified column references

## Test plan
- Verified query produces identical output hash (865c23a422cbcf2a3b53ea4d94c38ab656f5e8357e459b1c35bb2d8e9d26a559)
- Reduced average execution time:
```
+------+----------+-----------+-------------+---------------+-----------+------------+-------------+--------------+
| Test | Avg Time | Base Time | Time Change | % Time Change |  Avg Mem  |  Base Mem  |  Mem Change | % Mem Change |
+------+----------+-----------+-------------+---------------+-----------+------------+-------------+--------------+
|  0   |   596    |    2517   |    -1921    |      -76      | 126231182 | 3635406450 | -3509175268 |     -97      |
|  1   |   1004   |    2534   |    -1530    |      -60      | 125106384 | 3686434480 | -3561328096 |     -97      |
+------+----------+-----------+-------------+---------------+-----------+------------+-------------+--------------+
```